LPD-29126 Move MailMessageListener and MailEngine to the module
