+++ modules/apps/mail/mail-messaging-impl/src/main/java/com/liferay/mail/messaging/internal/MailEngine.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.mail.internal;
+package com.liferay.mail.messaging.internal;
 
 import com.liferay.mail.kernel.model.Account;
 import com.liferay.mail.kernel.model.FileAttachment;
