+++ modules/apps/portal/portal-util-test/src/testIntegration/java/com/liferay/portal/util/mail/test/MailEngineTest.java
@@ -7,8 +7,8 @@ package com.liferay.portal.util.mail.test;
 
 import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
 import com.liferay.mail.kernel.model.MailMessage;
-import com.liferay.mail.kernel.service.MailService;
 import com.liferay.petra.reflect.ReflectionUtil;
+import com.liferay.portal.kernel.module.util.BundleUtil;
 import com.liferay.portal.kernel.test.ReloadURLClassLoader;
 import com.liferay.portal.kernel.test.rule.AggregateTestRule;
 import com.liferay.portal.kernel.util.GetterUtil;
@@ -36,6 +36,10 @@ import org.junit.ClassRule;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
+
+import org.osgi.framework.Bundle;
+import org.osgi.framework.FrameworkUtil;
+import org.osgi.framework.wiring.BundleWiring;
 
 /**
  * @author Manuel de la Peña
@@ -178,12 +182,18 @@ public class MailEngineTest {
 				PropsKeys.DATA_LIMIT_MAIL_MESSAGE_MAX_PERIOD,
 				String.valueOf(maxMailMessagePeriod));
 
-			Class<?> clazz = MailService.class;
+			Bundle curBundle = FrameworkUtil.getBundle(MailEngineTest.class);
 
-			ClassLoader classLoader = clazz.getClassLoader();
+			Bundle bundle = BundleUtil.getBundle(
+				curBundle.getBundleContext(),
+				"com.liferay.mail.messaging.impl");
+
+			BundleWiring bundleWiring = bundle.adapt(BundleWiring.class);
+
+			ClassLoader classLoader = bundleWiring.getClassLoader();
 
 			Class<?> mailEngineClass = classLoader.loadClass(
-				"com.liferay.mail.internal.MailEngine");
+				"com.liferay.mail.messaging.internal.MailEngine");
 
 			Class<?> innerClass = mailEngineClass.getDeclaredClasses()[0];
